requirements.txt
dj-database-url==0.3.0
Django==1.8.5
djangorestframework==3.3.1
Markdown==2.6.3
psycopg2==2.6.1
pydash==3.4.3
Pygments==2.0.2
